extract Setup and Cleanup folders builder to a separate code
