+++ postman-generator/src/main/scala/generator/PredefinedCollectionItems.scala
@@ -2,8 +2,22 @@ package generator
 
 import generator.Heuristics.PathVariable
 import io.flow.postman.v0.models._
+import io.flow.postman.v0.{models => postman}
 
 object PredefinedCollectionItems {
+
+  def addItemTests(item: postman.Item): postman.Item = {
+
+    val method = item.request.method.getOrElse("").toString.toUpperCase
+    if (Seq("GET", "PUT", "POST").contains(method)) {
+      val test = PredefinedCollectionItems.testEventResponseStatusOk(
+        f"$method requests should return 2xx"
+      )
+      item.copy(event = Option(item.event.toSeq.flatten :+ test))
+    } else {
+      item
+    }
+  }
 
   def testEventResponseStatusOk(testTitle: String): Event = {
     Event(
